requirements.txt
asgiref
Django
django-cors-headers
django-debug-toolbar
django-filter
djangorestframework
pytz
sqlparse
psycopg2-binary
python-dotenv
django-allauth
dj-rest-auth
social-auth-core
google-auth 
google-auth-oauthlib
google-auth-httplib2
python-decouple
django-debug-toolbar
stripe